[KYUUBI #1268] Simplify extension code with Spark3.2 and Spark3.1 by unify zorder code
<!--
Thanks for sending a pull request!

Here are some tips for you:
  1. If this is your first time, please read our contributor guidelines: https://kyuubi.readthedocs.io/en/latest/community/contributions.html
  2. If the PR is related to an issue in https://github.com/apache/incubator-kyuubi/issues, add '[KYUUBI #XXXX]' in your PR title, e.g., '[KYUUBI #XXXX] Your PR title ...'.
  3. If the PR is unfinished, add '[WIP]' in your PR title, e.g., '[WIP][KYUUBI #XXXX] Your PR title ...'.
-->

### _Why are the changes needed?_
<!--
Please clarify why the changes are needed. For instance,
  1. If you add a feature, you can talk about the use case of it.
  2. If you fix a bug, you can clarify why it is a bug.
-->
Reduce file and code size to help maintain the extension.

- move all zorder code into common module
- add KyuubiSparkSQLCommonExtension in common module

### _How was this patch tested?_
Psss CI

Closes #1268 from ulysses-you/simplify-extension.

Closes #1268

0eeb7161 [ulysses-you] add link
c4d35c5c [ulysses-you] simplify

Authored-by: ulysses-you <ulyssesyou18@gmail.com>
