+++ dev/kyuubi-extension-spark-common/src/main/scala/org/apache/kyuubi/sql/zorder/OptimizeZorderStatementBase.scala
@@ -30,3 +30,14 @@ abstract class OptimizeZorderStatementBase extends UnaryNode {
   override def child: LogicalPlan = query
   override def output: Seq[Attribute] = child.output
 }
+
+/**
+ * A zorder statement that contains we parsed from SQL.
+ * We should convert this plan to certain command at Analyzer.
+ */
+case class OptimizeZorderStatement(
+    tableIdentifier: Seq[String],
+    query: LogicalPlan) extends OptimizeZorderStatementBase {
+  protected def withNewChildInternal(newChild: LogicalPlan): LogicalPlan =
+    copy(query = newChild)
+}
